refactor: 수동 로또를 받아오는 방식 수정

+++ src/main/kotlin/lotto/model/WinningLottoCalculator.kt
@@ -26,8 +26,8 @@ class WinningLottoCalculator(private val winningNumbers: Lotto, private val bonu
         purchaseAmount: Int,
         winningResult: Map<WinningRank, Int>,
     ): Double {
-        val profitAmount = calculateProfitAmount(winningResult)
-        return (profitAmount / purchaseAmount).toDouble()
+        val profitAmount = calculateProfitAmount(winningResult).toDouble()
+        return profitAmount / purchaseAmount.toDouble()
     }
 
     companion object {
